Extract makeCachedRequest to a helper module
